+++ source/config/configFlags.py
@@ -310,11 +310,11 @@ class ReportNotSupportedLanguage(DisplayStringStrEnum):
 	def _displayStringLabels(self) -> dict["ReportNotSupportedLanguage", str]:
 		return {
 			# Translators: A label for an option to report when the language of the text being read is not supported by the current synthesizer.
-			self.SPEECH: pgettext("languageNotSupported", "Speech"),
+			self.SPEECH: pgettext("reportLanguage", "Speech"),
 			# Translators: A label for an option to report when the language of the text being read is not supported by the current synthesizer.
-			self.BEEP: pgettext("languageNotSupported", "Beep"),
+			self.BEEP: pgettext("reportLanguage", "Beep"),
 			# Translators: A label for an option to report when the language of the text being read is not supported by the current synthesizer.
-			self.OFF: pgettext("languageNotSupported", "Off"),
+			self.OFF: pgettext("reportLanguage", "Off"),
 		}